perf(client): increased performance on tick
In the previous version there were three threads one running every 1000 ticks, one every tick and the other one running every 100 ticks. I've moved the loading of the minimap to the resource start. Since the removal of every component of the old map works good even every 1000 ticks, I've moved that section to when the player is in a vehicle. We've gone from 0.04/5ms to 0.01/2ms in idle. That's big performance achievement.